cleanup lucene dependency
