+++ kifi-backend/modules/common/app/com/keepit/scraper/Signature.scala
@@ -1,24 +1,8 @@
 package com.keepit.scraper
 
-import org.apache.lucene.analysis.standard.StandardTokenizer
-import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
-import org.apache.lucene.util.Version
-import java.io.StringReader
 import java.util.Arrays
 import javax.xml.bind.DatatypeConverter._
 import scala.math._
-
-object Signature {
-
-  import play.api.libs.json._
-  import play.api.libs.functional.syntax._
-
-  implicit val format: Format[Signature] = Format[Signature](Reads.of[String].fmap(Signature.apply), Writes { signature => JsString(signature.toBase64()) })
-
-  def apply(arr: Array[Byte]) = new Signature(arr)
-  def apply(base64: String) = new Signature(parseBase64Binary(base64))
-  def apply(fields: Seq[String]): Signature = fields.foldLeft(new SignatureBuilder) { (builder, text) => builder.add(text) }.build
-}
 
 class Signature(val bytes: Array[Byte]) {
 
@@ -41,68 +25,83 @@ class Signature(val bytes: Array[Byte]) {
   override def toString() = classOf[Signature].getName() + bytes.mkString("(", ",", ")")
 }
 
-class SignatureBuilder(windowSize: Int = 20) {
+object Signature {
 
-  private[this] val sketch = new Array[Int](100)
-  Arrays.fill(sketch, Int.MaxValue)
+  import play.api.libs.json._
+  import play.api.libs.functional.syntax._
 
-  private[this] val window = new Array[Int](windowSize + 1)
-  Arrays.fill(sketch, Int.MaxValue)
+  implicit val format: Format[Signature] = Format[Signature](Reads.of[String].fmap(Signature.apply), Writes { signature => JsString(signature.toBase64()) })
 
-  private[this] val cancelerShift = windowSize % 32
+  val SIZE = 100
+  def apply(arr: Array[Byte]) = new Signature(arr)
+  def apply(base64: String) = new Signature(parseBase64Binary(base64))
 
-  private[this] var ptr = windowSize
-  private[this] var canceler = 0
+  def empty(): Signature = {
+    val bytes = new Array[Byte](Signature.SIZE)
+    Arrays.fill(bytes, Int.MaxValue.toByte)
+    Signature(bytes)
+  }
 
-  def add(text: String) = {
-    val ts = new StandardTokenizer(Version.LUCENE_47, new StringReader(text))
-    val termAttr = ts.getAttribute(classOf[CharTermAttribute])
-    var h = window(ptr % windowSize)
+  abstract class Builder(windowSize: Int) {
 
-    try {
-      ts.reset()
-      while (ts.incrementToken()) {
-        h = ((h >>> 31) | (h << 1)) ^ hash(termAttr.buffer(), termAttr.length())
+    private[this] val sketch = new Array[Int](Signature.SIZE)
+    Arrays.fill(sketch, Int.MaxValue)
+
+    private[this] val window = new Array[Int](windowSize + 1)
+    Arrays.fill(sketch, Int.MaxValue)
+
+    private[this] val cancelerShift = windowSize % 32
+
+    private[this] var ptr = windowSize
+    private[this] var canceler = 0
+
+    def add(text: String): Builder = {
+      var h = window(ptr % windowSize)
+      tokenize(text) { (termBuf: Array[Char], termLen: Int) =>
+        h = ((h >>> 31) | (h << 1)) ^ hash(termBuf, termLen)
         ptr += 1
         canceler = window(ptr % windowSize)
         window(ptr % windowSize) = h
         updateSketch(h ^ ((canceler << cancelerShift) | (canceler >>> (32 - cancelerShift))))
-        ptr
       }
-      ts.end()
-    } finally {
-      ts.close()
+      this
     }
-    this
-  }
 
-  def build() = {
-    // borrowing the idea from b-bit minwise hash. we take the lowest 8 bits to save space.
-    Signature(sketch.map { _.toByte }.toArray)
-  }
+    protected def tokenize(text: String)(addTerm: (Array[Char], Int) => Unit): Unit
 
-  private def hash(arr: Array[Char], len: Int) = {
-    // Adler-32
-    var a = 0
-    var b = 0
-    var i = 0
-    while (i < len) {
-      val c = arr(i)
-      a += c
-      b += b + c
-      i += 1
+    def add(fields: Seq[String]): Builder = {
+      fields.foreach(add)
+      this
     }
-    ((a % 65521) + (b % 65521) * 65536)
-  }
 
-  private def updateSketch(seed: Int) {
-    var v = seed.toLong & 0x7FFFFFFFFFFFFFFFL
-    var i = 0
-    val sketchSize = sketch.length
-    while (i < 100) {
-      v = (v * 0x5DEECE66DL + 0x123456789L) & 0x7FFFFFFFFFFFFFFFL // linear congruential generator
-      sketch(i) = min(sketch(i), (v % 0x7FFFFFFFL).toInt) // positive int
-      i += 1
+    def build() = {
+      // borrowing the idea from b-bit minwise hash. we take the lowest 8 bits to save space.
+      Signature(sketch.map { _.toByte }.toArray)
+    }
+
+    private def hash(arr: Array[Char], len: Int) = {
+      // Adler-32
+      var a = 0
+      var b = 0
+      var i = 0
+      while (i < len) {
+        val c = arr(i)
+        a += c
+        b += b + c
+        i += 1
+      }
+      ((a % 65521) + (b % 65521) * 65536)
+    }
+
+    private def updateSketch(seed: Int) {
+      var v = seed.toLong & 0x7FFFFFFFFFFFFFFFL
+      var i = 0
+      val sketchSize = sketch.length
+      while (i < Signature.SIZE) {
+        v = (v * 0x5DEECE66DL + 0x123456789L) & 0x7FFFFFFFFFFFFFFFL // linear congruential generator
+        sketch(i) = min(sketch(i), (v % 0x7FFFFFFFL).toInt) // positive int
+        i += 1
+      }
     }
   }
 }
