+++ kifi-backend/modules/search/app/com/keepit/search/user/UserIndexer.scala
@@ -1,10 +1,10 @@
 package com.keepit.search.user
 
-import com.keepit.common.db.Id
-import com.keepit.common.db.SequenceNumber
+import com.keepit.common.db.{ ExternalId, Id, SequenceNumber, State }
 import com.keepit.model.User
 import com.keepit.social.BasicUser
 import com.keepit.search.index.{ IndexDirectory, Indexable, Indexer, DefaultAnalyzer }
+import org.apache.lucene.store.{ InputStreamDataInput, OutputStreamDataOutput }
 import org.apache.lucene.util.Version
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.shoebox.ShoeboxServiceClient
@@ -12,12 +12,12 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.concurrent.Future
-import com.keepit.common.db.State
 import com.keepit.model.UserStates._
 import com.keepit.model.ExperimentType
 import com.keepit.search.IndexInfo
 import com.keepit.typeahead.PrefixFilter
 import com.keepit.common.mail.EmailAddress
+import java.io.{ ByteArrayInputStream, ByteArrayOutputStream }
 
 object UserIndexer {
   val luceneVersion = Version.LUCENE_47
@@ -30,6 +30,37 @@ object UserIndexer {
   val PREFIX_MAX_LEN = 8 // do not change this number unless you do reindexing immediately
 
   val toBeDeletedStates = Set[State[User]](INACTIVE, PENDING, BLOCKED, INCOMPLETE_SIGNUP)
+}
+
+object BasicUserSerializer {
+  def toByteArray(basicUser: BasicUser): Array[Byte] = {
+    val bos = new ByteArrayOutputStream()
+    val oos = new OutputStreamDataOutput(bos)
+    oos.writeByte(1) // version
+    oos.writeString(basicUser.externalId.toString)
+    oos.writeString(basicUser.firstName)
+    oos.writeString(basicUser.lastName)
+    oos.writeString(basicUser.pictureName)
+    oos.close()
+    bos.close()
+    bos.toByteArray()
+  }
+
+  def fromByteArray(bytes: Array[Byte], offset: Int, length: Int): BasicUser = {
+    val in = new InputStreamDataInput(new ByteArrayInputStream(bytes, offset, length))
+
+    val version = in.readByte().toInt
+    if (version != 1) {
+      throw new Exception(s"invalid data [version=${version}]")
+    }
+
+    BasicUser(
+      externalId = ExternalId[User](in.readString),
+      firstName = in.readString,
+      lastName = in.readString,
+      pictureName = in.readString
+    )
+  }
 }
 
 class UserIndexer(
@@ -133,7 +164,7 @@ class UserIndexer(
       val expField = buildIteratorField[String](USER_EXPERIMENTS, experiments.map { _.value }.toIterator)(x => x)
       doc.add(expField)
 
-      val basicUserField = buildBinaryDocValuesField(BASIC_USER_FIELD, BasicUser.toByteArray(basicUser))
+      val basicUserField = buildBinaryDocValuesField(BASIC_USER_FIELD, BasicUserSerializer.toByteArray(basicUser))
       doc.add(basicUserField)
 
       val prefixField = buildIteratorField[String](PREFIX_FIELD, genPrefix(user).toIterator)(x => x)
